backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ainovel-backend"
version = "0.1.0"
description = "Snowflake Engine backend (Phase 1)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "pydantic>=2.7.0",
  "instructor>=1.3.3",
  "google-generativeai>=0.6.0",
  "uvicorn[standard]>=0.30.0",
  "kuzu>=0.11.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pytest-mock>=3.14.0",
  "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.editable]
packages = ["app"]
